requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.89.1
h11==0.14.0
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
pydantic==1.10.4
python-dotenv==0.21.0
requests==2.28.2
rsa==4.9
sniffio==1.3.0
starlette==0.22.0
Telethon==1.26.1
typing_extensions==4.4.0
urllib3==1.26.14
uuid==1.30
uvicorn==0.20.0
